Display terms in a scrollable block and a form to consent to them. Actual consent request was changed to require the POST method. Some renaming.

+++ webapp/byceps/blueprints/terms/views.py
@@ -7,7 +7,7 @@ byceps.blueprints.terms.views
 :Copyright: 2006-2015 Jochen Kupperschmidt
 """
 
-from flask import abort, redirect
+from flask import abort, redirect, request
 
 from ...database import db
 from ...util.framework import create_blueprint, flash_error, flash_success
@@ -15,6 +15,7 @@ from ...util.templating import templated
 
 from ..verification_token import service as verification_token_service
 
+from .forms import ConsentForm
 from .models import Version
 from . import service
 
@@ -30,16 +31,39 @@ def view_current():
     return {'version': version}
 
 
-@blueprint.route('/<uuid:version_id>/accept/<uuid:token>')
+@blueprint.route('/<uuid:version_id>/consent/<uuid:token>')
 @templated
-def accept(version_id, token):
-    """Accept that version of the terms."""
+def consent_form(version_id, token, *, erroneous_form=None):
+    """Show that version of the terms, and a form to consent to it."""
     version = Version.query.get_or_404(version_id)
 
     verification_token = verification_token_service.find_for_terms_consent_by_token(token)
     if verification_token is None:
         flash_error('Unbekannter Bestätigungscode.')
         abort(404)
+
+    form = erroneous_form if erroneous_form else ConsentForm()
+
+    return {
+        'version': version,
+        'token': token,
+        'form': form,
+    }
+
+
+@blueprint.route('/<uuid:version_id>/consent/<uuid:token>', methods=['POST'])
+def consent(version_id, token):
+    """Consent to that version of the terms."""
+    version = Version.query.get_or_404(version_id)
+
+    verification_token = verification_token_service.find_for_terms_consent_by_token(token)
+    if verification_token is None:
+        flash_error('Unbekannter Bestätigungscode.')
+        abort(404)
+
+    form = ConsentForm(request.form)
+    if not form.validate():
+        return consent_form(version_id, token, erroneous_form=form)
 
     user = verification_token.user
     db.session.delete(verification_token)
